+++ performance/hello/main.dart
@@ -1,13 +1,14 @@
 /// A basic server that prints "Hello, world!"
 library performance.hello;
 
+import 'dart:async';
 import 'dart:io';
 import 'dart:isolate';
 import 'package:angel_framework/angel_framework.dart';
 import 'package:angel_framework/metrics.dart';
 
 main() {
-  for (int i = 0; i < Platform.numberOfProcessors - 1; i++) {
+  for (int i = 0; i < Platform.numberOfProcessors; i++) {
     Isolate.spawn(start, i + 1);
   }